Split the DNS provider into core/dns/coredns_provider
Everything specific to the router serving DNS itself now lives in one
subpackage, so what is left in core/dns/ is provider-agnostic and would be
unchanged if the router stopped being a provider:

  service_api.py       the contract both sides share
  records.py           what a record is
  client.py            calling the service
  dynamic_dns.py       the public-IP watcher (renamed from dynamic.py)
  public_ip.py         detect and store where the instance is
  coredns_provider/    service.py, coredns.py, zonefile.py, templates

Service identity and the grant type move out of the provider's service.py into
service_api.py, so the client and the provider agree on what a grant means
without either importing the other. client.py reaches into the subpackage for
one thing only: dispatching in-process rather than making the router talk to
itself over loopback.

router_grants now returns typed Grants instead of raw permission dicts; only
the remote path serializes them, so the in-process call no longer round-trips
through JSON to be parsed straight back.

Test modules renamed to match what they cover.
