+++ compute_space/src/compute_space/tests/test_coredns.py
@@ -7,13 +7,13 @@ from typing import Any
 
 import pytest
 
-import compute_space.core.dns.coredns as dns_mod
-import compute_space.core.dns.zonefile as zonefile_mod
+import compute_space.core.dns.coredns_provider.coredns as dns_mod
+import compute_space.core.dns.coredns_provider.zonefile as zonefile_mod
 from compute_space.config import DefaultConfig
-from compute_space.core.dns.coredns import DnsZone
-from compute_space.core.dns.coredns import public_dns_zones
-from compute_space.core.dns.coredns import reload_coredns_for_domains
-from compute_space.core.dns.coredns import set_active_coredns
+from compute_space.core.dns.coredns_provider.coredns import DnsZone
+from compute_space.core.dns.coredns_provider.coredns import public_dns_zones
+from compute_space.core.dns.coredns_provider.coredns import reload_coredns_for_domains
+from compute_space.core.dns.coredns_provider.coredns import set_active_coredns
 from compute_space.core.dns.records import DnsRecord
 from compute_space.core.domains import Domain
 from compute_space.core.domains import DomainRecord
