+++ compute_space/src/compute_space/tests/test_dynamic_dns.py
@@ -10,9 +10,9 @@ import httpx
 import pytest
 
 from compute_space.config import DefaultConfig
-from compute_space.core.dns import dynamic
+from compute_space.core.dns import dynamic_dns as dynamic
 from compute_space.core.dns import public_ip as public_ip_mod
-from compute_space.core.dns import zonefile
+from compute_space.core.dns.coredns_provider import zonefile
 from compute_space.core.dns.public_ip import detect_public_ip
 from compute_space.core.dns.public_ip import effective_public_ip
 from compute_space.core.dns.public_ip import is_public_ipv4
