+++ compute_space/src/compute_space/web/start.py
@@ -25,12 +25,12 @@ from compute_space.core.caddy import start_caddy
 from compute_space.core.caddy import unix_admin_address
 from compute_space.core.containers import CONTAINER_GATEWAY_IP
 from compute_space.core.dns.client import uses_local_dns
-from compute_space.core.dns.coredns import CoreDnsProcess
-from compute_space.core.dns.coredns import coredns_is_needed
-from compute_space.core.dns.coredns import public_dns_zones
-from compute_space.core.dns.coredns import set_active_coredns
-from compute_space.core.dns.coredns import start_coredns
-from compute_space.core.dns.dynamic import start_dynamic_dns_thread
+from compute_space.core.dns.coredns_provider.coredns import CoreDnsProcess
+from compute_space.core.dns.coredns_provider.coredns import coredns_is_needed
+from compute_space.core.dns.coredns_provider.coredns import public_dns_zones
+from compute_space.core.dns.coredns_provider.coredns import set_active_coredns
+from compute_space.core.dns.coredns_provider.coredns import start_coredns
+from compute_space.core.dns.dynamic_dns import start_dynamic_dns_thread
 from compute_space.core.dns.public_ip import effective_public_ip
 from compute_space.core.dns.public_ip import seed_public_ip
 from compute_space.core.domains import Domain
